Add files via upload
HunterHolsteadmain.c is an example of how to run all of the corresponding function tests with the .h file. Setters have been written for the clients in the current state however future versions will see the database functions creating structs that hold information in variables for easier access from getter methods. The overall flow of the client program will be reworked as well.